+++ actions/check-licenses/src/licenses.ts
@@ -23,6 +23,7 @@ export const ALLOWED_LICENSES = [
   'ISC',
   'MIT',
   'Miros',
+  'MPL-2.0',
   'Multics',
   'Naumen',
   'NTP',