+++ actions/check-licenses/src/npmPackages.ts
@@ -1,1 +1,1 @@
-export const EXCLUDE_NPM_PACKAGES = ['webpack', '@types', '@babel', 'k6'];
+export const EXCLUDE_NPM_PACKAGES = ['webpack', '@types', '@babel', 'k6', '@nuxt/ui-templates'];