+++ packages/middleware/__tests__/unit/handlers/getApiFunction.spec.ts
@@ -12,41 +12,42 @@ describe("getApiFunction", () => {
     },
   };
 
-  it("should return the correct API function when extensionName is provided", () => {
+  it("should return the correct API function when extensionName is provided", async () => {
     const functionName = "function1";
     const extensionName = "extension1";
     const expectedApiFn = apiClient.api[extensionName][functionName];
 
-    const apiFn = getApiFunction(apiClient, functionName, extensionName);
+    const apiFn = await getApiFunction(apiClient, functionName, extensionName);
 
     expect(apiFn).toEqual(expectedApiFn);
   });
 
-  it("should return the correct API function when extensionName is not provided", () => {
+  it("should return the correct API function when extensionName is not provided", async () => {
     const functionName = "function3";
     const expectedApiFn = apiClient.api[functionName];
 
-    const apiFn = getApiFunction(apiClient, functionName);
+    const apiFn = await getApiFunction(apiClient, functionName);
 
     expect(apiFn).toEqual(expectedApiFn);
   });
 
-  it("should throw an error when the API function is not found with extensionName", () => {
+  it("should throw an error when the API function is not found with extensionName", async () => {
     const functionName = "function5";
     const extensionName = "extension1";
 
-    expect(() => {
-      getApiFunction(apiClient, functionName, extensionName);
-    }).toThrowError(
-      `Extension "${extensionName}" is not namespaced or the function "${functionName}" is not available in the namespace.`
+    await expect(async () => {
+      await getApiFunction(apiClient, functionName, extensionName);
+    }).rejects.toThrowError(
+      `Failed to resolve apiClient or function: Extension "${extensionName}" is not namespaced or the function "${functionName}" is not available in the namespace.`
     );
   });
 
-  it("should throw an error when the API function is not found without extensionName", () => {
+  it("should throw an error when the API function is not found without extensionName", async () => {
     const functionName = "function5";
-
-    expect(() => {
-      getApiFunction(apiClient, functionName);
-    }).toThrowError(`The function "${functionName}" is not registered.`);
+    await expect(async () => {
+      await getApiFunction(apiClient, functionName);
+    }).rejects.toThrowError(
+      `Failed to resolve apiClient or function: The function "${functionName}" is not registered.`
+    );
   });
 });