+++ packages/next/__tests__/apps/app-router/sdk/sdk.server.ts
@@ -1,0 +1,9 @@
+import { createSdk } from "@vue-storefront/next";
+import { getSdkOptions } from "./options";
+import { getSdkConfig } from "./config";
+
+export const { getSdk } = createSdk(
+  getSdkOptions(), getSdkConfig()
+);
+
+export type Sdk = ReturnType<typeof getSdk>;