+++ packages/nuxt/__tests__/app/sdk.config.ts
@@ -1,0 +1,12 @@
+import type { Endpoints } from "@storefront/shared";
+
+export default defineSdkConfig(
+  ({ buildModule, middlewareModule, getRequestHeaders, config }) => {
+    return {
+      example: buildModule(middlewareModule<Endpoints>, {
+        apiUrl: `${config.middlewareUrl}/test_integration`,
+        defaultRequestConfig: { headers: { ...getRequestHeaders() } },
+      }),
+    };
+  }
+);