+++ packages/nuxt/src/runtime/utils/defaults.ts
@@ -1,0 +1,18 @@
+export function getDefaultMethodsRequestConfig() {
+  return {
+    unifiedCommerce: {
+      middlewareModule: {
+        getCategories: { method: "GET" },
+        getProductDetails: { method: "GET" },
+        getProductReviews: { method: "GET" },
+        getProducts: { method: "GET" },
+        searchProducts: { method: "GET" },
+      },
+    },
+    unifiedCms: {
+      middlewareModule: {
+        getEntries: { method: "GET" },
+      },
+    },
+  } as const;
+}