+++ packages/sdk/src/modules/middlewareModule/utils/getRequestSender.ts
@@ -17,7 +17,13 @@ import { SdkHttpError } from "./SdkHttpError";
  * handling errors, and executing HTTP requests.
  */
 export const getRequestSender = (options: Options): RequestSender => {
-  const { apiUrl, ssrApiUrl, defaultRequestConfig = {} } = options;
+  const {
+    apiUrl,
+    ssrApiUrl,
+    defaultRequestConfig = {},
+    methodsRequestConfig = {},
+    cdnCacheBustingId = undefined,
+  } = options;
 
   const getUrl = (
     path: string,
@@ -40,13 +46,21 @@ export const getRequestSender = (options: Options): RequestSender => {
     }
 
     // If there are query params, append them to the URL as `?body=[<strignified query params>]`
-    const serializedParams = encodeURIComponent(JSON.stringify(params));
+    const serializedBody = encodeURIComponent(JSON.stringify(params));
+    // Serialize CDN cache busting ID
+    const serializedCdnCacheBustingId = cdnCacheBustingId
+      ? `&cdnCacheBustingId=${encodeURIComponent(cdnCacheBustingId)}`
+      : "";
 
-    return `${url}?body=${serializedParams}`;
+    return `${url}?body=${serializedBody}${serializedCdnCacheBustingId}`;
   };
 
-  const getConfig = (config: RequestConfig): ComputedConfig => {
+  const getConfig = (
+    config: RequestConfig,
+    methodConfig: RequestConfig
+  ): ComputedConfig => {
     const { method, headers } = config;
+    const { headers: methodHeaders = {} } = methodConfig;
     const defaultHeaders = {
       "Content-Type": "application/json",
       Accept: "application/json",
@@ -54,6 +68,7 @@ export const getRequestSender = (options: Options): RequestSender => {
     };
     const mergedHeaders = {
       ...defaultHeaders,
+      ...methodHeaders,
       ...headers,
     };
 
@@ -71,14 +86,10 @@ export const getRequestSender = (options: Options): RequestSender => {
     };
   };
 
-  const defaultHTTPClient: HTTPClient = async (
-    url: string,
-    params: unknown[],
-    config?: ComputedConfig
-  ) => {
+  const defaultHTTPClient: HTTPClient = async (url, params, config) => {
     const response = await fetch(url, {
       ...config,
-      body: JSON.stringify(params),
+      body: config?.method === "GET" ? undefined : JSON.stringify(params),
       credentials: "include",
     });
 
@@ -103,19 +114,25 @@ export const getRequestSender = (options: Options): RequestSender => {
       httpClient = defaultHTTPClient,
       errorHandler = defaultErrorHandler,
     } = options;
-    const { method = "POST", headers = {}, ...restConfig } = config ?? {};
-    const computedParams = method === "GET" ? [] : params;
-    const finalUrl = getUrl(methodName, method, params);
-    const finalConfig = getConfig({ method, headers, ...restConfig });
+    const { method, headers = {}, ...restConfig } = config ?? {};
+    const methodConfig = methodsRequestConfig[methodName] || {};
+    const finalMethod =
+      method || methodConfig.method || defaultRequestConfig.method || "POST";
+    const finalUrl = getUrl(methodName, finalMethod, params);
+    const finalParams = finalMethod === "GET" ? [] : params;
+    const finalConfig = getConfig(
+      { method: finalMethod, headers, ...restConfig },
+      methodConfig
+    );
 
     try {
-      return await httpClient(finalUrl, computedParams, finalConfig);
+      return await httpClient(finalUrl, finalParams, finalConfig);
     } catch (error) {
       return await errorHandler({
         error,
         methodName,
         url: finalUrl,
-        params: computedParams,
+        params: finalParams,
         config: finalConfig,
         httpClient,
       });